Force "Retrieve list of existing network-scripts" to run even in check mode
This command is guaranteed to not cause any side effects but Ansible can't
know that so it skips the command in check mode. Here we tell Ansible it
is okay to run this command even in check mode, with the benefit that the
next tasks that depend on information gathered by this task, can also run
successfully.